% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_layer_CV.R
\name{inner_layer_CV}
\alias{inner_layer_CV}
\title{inner_layer_CV}
\usage{
inner_layer_CV(
  sum_stats,
  ped,
  G,
  bim,
  initial,
  num_snp,
  covar = NULL,
  fixG = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sum_stats}{dataframe including PRS weights from disease GWAS.}

\item{ped}{dataframe including phenotype and covariates.}

\item{G}{The genotype matrix of individual x SNP. Row names are individual ID and column names are SNP ID.}

\item{bim}{Allele info for the SNPs in genotype file. The following columns are required: SNP, Ref. Column names are needed.}

\item{initial}{Indicating the starting values of the predictive effects, 'PRS' or 'zero'.}

\item{num_snp}{number of total snps with non-zero effects.
It is used to determine the learning rate as
(factor * 1 / num_snp), where the factor is varied across a grid of values: 1, 10, 50, 70.}

\item{covar}{covariates to be adjusted in the model (other than Tr), such as "age,gender". This is optional. Default is NULL.}

\item{fixG}{logical, indicating whether the genetic main (G) effect is fixed. Default is FALSE.}

\item{verbose}{logical, indicating whether to print more details. Default is FALSE.}
}
\value{
A list of S_prog and S_pred values for the validation set, the training set, and the tuning parameters.
}
\description{
Inner layer of cross-validation.
}
\examples{
\dontrun{
data(sum_stats)
data(ped)
data(G)
data(bim)
result_list = inner_layer_CV(sum_stats = sum_stats, ped = ped, G = G, bim = bim, 
initial = "PRS", num_snp = 2349, covar = NULL, fixG = FALSE, verbose = TRUE)
}
}
