% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_tuning.R
\name{parameter_tuning}
\alias{parameter_tuning}
\title{parameter_tuning}
\usage{
parameter_tuning(
  X,
  s1_results = NULL,
  s1_files = NULL,
  s1_byCHR_files = NULL,
  fixG = FALSE,
  covar = NULL,
  verbose = FALSE
)
}
\arguments{
\item{X}{dataframe including phenotype and covariates.}

\item{s1_results}{a list of output from the "inner_layer_CV" function.}

\item{s1_files}{The file name for output from the "inner_layer_CV" function, if loading the result from a ".rds" file.}

\item{s1_byCHR_files}{a vector or list of file names for output from the "inner_layer_CV" function for each chromosome, if loading the results from ".rds" files.}

\item{fixG}{logical, indicating whether the genetic main (G) effect is fixed. Default is FALSE.}

\item{covar}{covariates to be adjusted in the model (other than Tr), such as "age,gender". This is optional. Default is NULL.}

\item{verbose}{logical, indicating whether to print more details. Default is FALSE.}
}
\description{
parameter tuning
}
\examples{
\dontrun{
data(ped)
data(inner_layer_CV_output)
result_list = parameter_tuning(X = ped, s1_results = inner_layer_CV_output)
}
}
