% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_algorithm.R
\name{rerun_algorithm}
\alias{rerun_algorithm}
\title{rerun_algorithm}
\usage{
rerun_algorithm(
  sum_stats,
  ped,
  G,
  bim,
  initial,
  num_snp,
  s2_results,
  best_criterion = "best_R2",
  covar = NULL,
  fixG = FALSE
)
}
\arguments{
\item{sum_stats}{dataframe including PRS weights from disease GWAS.}

\item{ped}{dataframe including phenotype and covariates.}

\item{G}{The genotype matrix of individual x SNP. Row names are individual ID and column names are SNP ID.}

\item{bim}{Allele info for the SNPs in genotype file. The following columns are required: SNP, Ref. Column names are needed.}

\item{initial}{Indicating the starting values of the predictive effects, 'PRS' or 'zero'.}

\item{num_snp}{number of total snps with non-zero effects.
It is used to determine the learning rate as
(factor * 1 / num_snp), where the factor is varied across a grid of values: 1, 10, 50, 70.}

\item{s2_results}{The output from the "parameter tuning".}

\item{best_criterion}{The criterion to maximize overall R2 ("best_R2") or the conditional R2 of GxT (conditional on G) ("best_condR2") when selecting the best parameters. Default is "best_R2".}

\item{covar}{covariates to be adjusted in the model (other than Tr), such as "age,gender". This is optional. Default is NULL.}

\item{fixG}{logical, indicating whether the genetic main (G) effect is fixed. Default is FALSE.}
}
\value{
A list of beta values and test scoresum values.
}
\description{
Rerun the algorithm and get the output with the best parameters
}
\examples{
\dontrun{
data(sum_stats)
data(ped)
data(G)
data(bim)
data(parameter_tuning_output)
result_list = rerun_algorithm(sum_stats = sum_stats, ped = ped, G = G, bim = bim, 
initial = "PRS", num_snp = 2349, s2_results = parameter_tuning_output)
}
}
