% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_PRS_over_CHR.R
\name{sum_PRS_over_CHR}
\alias{sum_PRS_over_CHR}
\title{sum_PRS_over_CHR}
\usage{
sum_PRS_over_CHR(s3_byCHR_files, fixG = FALSE)
}
\arguments{
\item{s3_byCHR_files}{a vector or list of file names for output from "rerun_algorithm" function for each chromosome}

\item{fixG}{logical, indicating whether the genetic main (G) effect is fixed. Default is FALSE.}
}
\value{
a data frame with the sum of PRS over all chromosomes for the testing samples
}
\description{
sum the PRS over all chromosomes for the testing samples
}
